# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Chat Atendimento API",
    "version": "1.0.0",
    "description": "\n        API REST para sistema de chat de atendimento em tempo real.\n        \n        ## Funcionalidades\n        - Autenticação JWT\n        - Chat em tempo real com WebSocket\n        - Upload de arquivos (AWS S3)\n        - Gerenciamento de atendentes\n        - Fila de atendimento\n        - Dashboard administrativo\n        - Histórico e análises\n        \n        ## Ambientes\n        - **Desenvolvimento**: http://localhost:5000\n        - **Staging**: http://52.90.17.204:5000\n        - **Produção**: https://api.brsi.com.br\n        \n        ## Autenticação\n        Todas as rotas (exceto login e register) requerem autenticação via JWT.\n        Envie o token no header: Authorization: Bearer {token}\n      ",
    "contact": {
      "name": "BR Sistemas",
      "email": "suporte@brsi.com.br",
      "url": "https://brsi.com.br"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://brsi.com.br/terms"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Servidor de Desenvolvimento (Local)"
    },
    {
      "url": "http://52.90.17.204:5000/api",
      "description": "Servidor de Staging (Testes)"
    },
    {
      "url": "https://api.brsi.com.br/api",
      "description": "Servidor de Produção"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Digite o token JWT obtido no login"
      }
    },
    "schemas": {
...
      "Chat": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID único da conversa",
            "example": "60d0fe4f5311236168a109cb"
          },
          "subject": {
            "type": "string",
            "description": "Assunto da conversa",
            "example": "Dúvida sobre produto"
          },
          "customer": {
            "type": "object",
            "description": "Dados do cliente",
            "properties": {
              "_id": {
                "type": "string"
              },
              "name": {
                "type": "string",
                "example": "Maria Santos"
              },
              "email": {
                "type": "string",
                "format": "email",
                "example": "maria@example.com"
              }
            }
          },
          "agent": {
            "type": "string",
            "description": "ID do agente atribuído",
            "example": "60d0fe4f5311236168a109ca"
          },
          "status": {
            "type": "string",
            "enum": [
              "waiting",
              "active",
              "closed"
            ],
            "description": "Status atual da conversa",
            "example": "active"
          },
          "department": {
            "type": "string",
            "description": "Departamento responsável",
            "example": "vendas"
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "description": "Avaliação do atendimento",
            "example": 5
          },
          "ratingComment": {
            "type": "string",
            "description": "Comentário da avaliação",
            "example": "Atendimento excelente!"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de criação"
          },
          "closedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de encerramento"
          },
          "lastActivity": {
            "type": "string",
            "format": "date-time",
            "description": "Última atividade na conversa"
          }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID único da mensagem"
          },
          "conversationId": {
            "type": "string",
            "description": "ID da conversa"
          },
          "sender": {
            "type": "string",
            "description": "ID do remetente"
          },
          "senderType": {
            "type": "string",
            "enum": [
              "customer",
              "agent",
              "system"
            ],
            "description": "Tipo do remetente"
          },
          "content": {
            "type": "string",
            "description": "Conteúdo da mensagem",
            "example": "Olá, como posso ajudar?"
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "file",
              "audio"
            ],
            "description": "Tipo de mensagem",
            "example": "text"
          },
          "file": {
            "type": "object",
            "description": "Arquivo anexado (se houver)",
            "properties": {
              "url": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "size": {
                "type": "number"
              },
              "mimetype": {
                "type": "string"
              }
            }
          },
          "readBy": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "string"
                },
                "readAt": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            },
            "description": "Usuários que leram a mensagem"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de envio"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/history/conversations": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "Buscar histórico de conversas",
        "description": "Retorna histórico de conversas com filtros opcionais",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "waiting",
                "active",
                "closed"
              ]
            },
            "description": "Filtrar por status"
          },
          {
            "in": "query",
            "name": "agentId",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por ID do agente"
          },
          {
            "in": "query",
            "name": "clientId",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por ID do cliente"
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Data inicial (formato YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Data final (formato YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Número da página"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Itens por página"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "closedAt",
                "rating"
              ],
              "default": "createdAt"
            },
            "description": "Campo para ordenação"
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            },
            "description": "Ordem de classificação"
          }
        ],
        "responses": {
          "200": {
            "description": "Histórico de conversas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "conversations": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Chat"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "pages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          }
        }
      }
    },
    "/api/history/conversations/{conversationId}": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "Obter detalhes da conversa",
        "description": "Retorna detalhes completos de uma conversa específica, incluindo mensagens",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "conversationId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da conversa"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes da conversa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "conversation": {
                      "$ref": "#/components/schemas/Chat"
                    },
                    "messages": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Message"
                      }
                    },
                    "attachments": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Conversa não encontrada"
          }
        }
      }
    },
    "/api/history/search": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "Buscar conversas",
        "description": "Busca conversas por texto nas mensagens ou metadados",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Termo de busca",
            "example": "problema no pagamento"
          },
          {
            "in": "query",
            "name": "searchIn",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "messages",
                  "subject",
                  "customer"
                ]
              },
              "default": [
                "messages",
                "subject"
              ]
            },
            "description": "Campos onde buscar"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resultados da busca",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "conversation": {
                            "$ref": "#/components/schemas/Chat"
                          },
                          "matches": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "message": {
                                  "type": "string"
                                },
                                "timestamp": {
                                  "type": "string",
                                  "format": "date-time"
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parâmetros inválidos"
          },
          "401": {
            "description": "Não autorizado"
          }
        }
      }
    },
    "/api/history/export/{conversationId}/csv": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "Exportar conversa em CSV",
        "description": "Exporta uma conversa e suas mensagens em formato CSV",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "conversationId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da conversa"
          }
        ],
        "responses": {
          "200": {
            "description": "Arquivo CSV",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string",
                  "example": "attachment; filename=\"conversa_60d0fe4f5311236168a109ca.csv\""
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Conversa não encontrada"
          }
        }
      }
    },
    "/api/history/export/{conversationId}/pdf": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "Exportar conversa em PDF",
        "description": "Exporta uma conversa e suas mensagens em formato PDF",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "conversationId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da conversa"
          }
        ],
        "responses": {
          "200": {
            "description": "Arquivo PDF",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string",
                  "example": "attachment; filename=\"conversa_60d0fe4f5311236168a109ca.pdf\""
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Conversa não encontrada"
          }
        }
      }
    },
    "/api/history/analytics": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "Obter análise de conversas",
        "description": "Retorna estatísticas e análises das conversas",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "period",
            "schema": {
              "type": "string",
              "enum": [
                "today",
                "week",
                "month",
                "year",
                "custom"
              ],
              "default": "month"
            },
            "description": "Período de análise"
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Data inicial (para período custom)"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Data final (para período custom)"
          },
          {
            "in": "query",
            "name": "groupBy",
            "schema": {
              "type": "string",
              "enum": [
                "day",
                "week",
                "month"
              ],
              "default": "day"
            },
            "description": "Agrupar dados por"
          }
        ],
        "responses": {
          "200": {
            "description": "Dados analíticos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "summary": {
                      "type": "object",
                      "properties": {
                        "totalConversations": {
                          "type": "integer"
                        },
                        "completedConversations": {
                          "type": "integer"
                        },
                        "avgResolutionTime": {
                          "type": "number",
                          "description": "Tempo médio de resolução em minutos"
                        },
                        "avgMessagesPerConversation": {
                          "type": "number"
                        },
                        "avgSatisfaction": {
                          "type": "number",
                          "minimum": 0,
                          "maximum": 5
                        }
                      }
                    },
                    "timeline": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "date": {
                            "type": "string",
                            "format": "date"
                          },
                          "conversations": {
                            "type": "integer"
                          },
                          "messages": {
                            "type": "integer"
                          },
                          "avgResponseTime": {
                            "type": "number"
                          }
                        }
                      }
                    },
                    "topAgents": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "agentId": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "totalConversations": {
                            "type": "integer"
                          },
                          "avgSatisfaction": {
                            "type": "number"
                          }
                        }
                      }
                    },
                    "peakHours": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "hour": {
                            "type": "integer"
                          },
                          "conversations": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Sem permissão - Apenas administradores"
          }
        }
      }
    },
...
  }
}